+++ internal/provider/vm_resource.go
@@ -162,6 +162,7 @@ func (r *vmResource) Schema(ctx context.Context, _ resource.SchemaRequest, resp 
 						"url":  schema.StringAttribute{Computed: true},
 						"addr": schema.StringAttribute{Computed: true},
 						"tcp":  schema.BoolAttribute{Computed: true},
+						"udp":  schema.BoolAttribute{Computed: true},
 					},
 				},
 			},
@@ -385,6 +386,7 @@ func refreshVMStatus(ctx context.Context, c *client.Client, m *vmResourceModel, 
 			URL:  types.StringValue(e.URL),
 			Addr: types.StringValue(e.Addr),
 			TCP:  types.BoolValue(e.TCP),
+			UDP:  types.BoolValue(e.UDP),
 		})
 	}
 	m.URL = types.StringValue(url)